test(ui): split chat follow-up E2E suite
